+++ src/com/madbros/tileminer/GameStates/FurnaceStateInput.java
@@ -80,11 +80,11 @@ public class FurnaceStateInput extends InventoryStateInput {
 	
 	public void additionalMouseMove() {
 		Rect mouseRect = new Rect(Helpers.getX(), Helpers.getY(), 1, 1);
-		Game.inventory.craftingMenu.handleMouseMove((int)mouseRect.x, (int)mouseRect.y);
+		Game.inventory.craftingMenu.handleMouseMove(mouseRect.x, mouseRect.y);
 		
 		FurnaceTile furnaceTile = (FurnaceTile) Game.level.activeBlocks[Game.inventory.currentInvActiveBlockX][Game.inventory.currentInvActiveBlockY].layers[OBJECT_LAYER];
 		if(furnaceTile.furnaceSlots[0].item.id != 0) {
-			furnaceTile.furnaceSlots[0].handleMouseMove((int)mouseRect.x, (int)mouseRect.y);
+			furnaceTile.furnaceSlots[0].handleMouseMove(mouseRect.x, mouseRect.y);
 		}
 		Slot[][] slots = {Game.inventory.invBar, Game.inventory.invBag, Game.inventory.craftingMenu.craftSlots, Game.inventory.invCrafted, Game.inventory.invClothing, furnaceTile.furnaceSlots, furnaceTile.fuelSlot, furnaceTile.craftedSlot};
 		for(int i = 0; i < slots.length; i++) {
